🧬 Refactored `Component` and folder structure
Why
- To improve the overall app Design and decouple unnecessary `Components` relations

How
- Removed the `group` CSS class (old artifact)
- Removed scrollbar commented styling (old artifact)
- Removed the `isBlank` property as `prop` from all the component
- Removed the CSS generated `background grid` in favour for generating `div` elements for each `Day` (broke `Event` generation 🐞)
- Decoupled all the unnecessary linking between components, mainly `Hours` and `Day` components
- Reworked the `Hours` and `Day` components code to accommodate for the unlinking
- Renamed components based on their function
- Restructured the whole project's folder structure
- Added the import of `subfolders` as JavaScript `modules`

Tags
 Changes to be committed:
	modified:   package-lock.json
	modified:   package.json
	renamed:    src/components/Day.vue -> src/components/Calendar/CalendarDay/CalendarDay.vue
	new file:   src/components/Calendar/CalendarDay/index.ts
	new file:   src/components/Calendar/CalendarHours/CalendarHours.vue
	new file:   src/components/Calendar/CalendarHours/index.ts
	renamed:    src/components/View.vue -> src/components/Calendar/CalendarView/CalendarView.vue
	new file:   src/components/Calendar/CalendarView/index.ts
	renamed:    src/components/DayHeader.vue -> src/components/Calendar/Headers/DayHeader/DayHeader.vue
	new file:   src/components/Calendar/Headers/DayHeader/index.ts
	renamed:    src/components/EventHeader.vue -> src/components/Calendar/Headers/EventsHeader/EventsHeader.vue
	new file:   src/components/Calendar/Headers/EventsHeader/index.ts
	deleted:    src/components/Hours.vue
	renamed:    src/components/Sidebar.vue -> src/components/Sidebar/Sidebar.vue
	new file:   src/components/Sidebar/index.ts
	new file:   src/components/package.json
	modified:   src/views/Calendar.vue
