+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "calendar-app-vue",
+  "name": "cl",
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
@@ -10,6 +10,7 @@
     "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false"
   },
   "dependencies": {
+    "@cl/components": "file:src/components",
     "pinia": "^2.0.33",
     "vue": "^3.2.47"
   },
