+++ src/components/Calendar/CalendarDay/index.ts
@@ -1,0 +1,2 @@
+import CalendarDay from './CalendarDay.vue';
+export default CalendarDay;
