+++ src/components/Calendar/CalendarHours/index.ts
@@ -1,0 +1,2 @@
+import CalendarHours from './CalendarHours.vue';
+export default CalendarHours;
