+++ src/components/Calendar/CalendarView/CalendarView.vue
@@ -2,14 +2,8 @@
   <div
     class="relative col-start-3 col-end-[12] row-start-3 row-end-[12] grid grid-cols-8 overflow-x-hidden overflow-y-scroll"
   >
-    <div class="h-full">
-      <Day isBlank class="z-0">
-        <template #hours>
-          <Hours v-for="hour in hours" :hour="hour" :key="hour" />
-        </template>
-      </Day>
-    </div>
-    <Day
+    <CalendarHours class="z-0" />
+    <CalendarDay
       class="z-0"
       v-for="(day, index) in weekDates"
       :key="day"
@@ -21,12 +15,11 @@
 
 <script lang="ts" setup>
 import { computed } from 'vue';
+import CalendarDay from '@cl/components/Calendar/CalendarDay';
+import CalendarHours from '@cl/components/Calendar/CalendarHours';
 import { useCalendarStore } from '@/stores/calendarStore';
-import Day from './Day.vue';
-import Hours from './Hours.vue';
 import { convertToStringDates } from '@/utils/Dates';
 
-const hours = 23;
 const store = useCalendarStore();
 
 const weekDates = computed(() => convertToStringDates(store.view, true));
@@ -40,10 +33,6 @@ const shortWeekDates = computed(() =>
   width: 6px;
   background-color: transparent;
 }
-
-/* *::-webkit-scrollbar-track {
-  background: orange;
-} */
 
 *::-webkit-scrollbar-thumb {
   background-color: rgb(229, 231, 235);
