+++ src/components/Calendar/CalendarView/index.ts
@@ -1,0 +1,2 @@
+import CalendarView from './CalendarView.vue';
+export default CalendarView;
