+++ src/components/Calendar/Headers/DayHeader/index.ts
@@ -1,0 +1,2 @@
+import DayHeader from './DayHeader.vue';
+export default DayHeader;
