+++ src/components/Calendar/Headers/EventsHeader/index.ts
@@ -1,0 +1,2 @@
+import EventsHeader from './EventsHeader.vue';
+export default EventsHeader;
