+++ src/components/Sidebar/index.ts
@@ -1,0 +1,2 @@
+import Sidebar from './Sidebar.vue';
+export default Sidebar;
