+++ src/components/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@cl/components",
+  "version": "0.0.0",
+  "main": "index.ts"
+}
